requirements.txt
pandas
streamlit
altair
vega_datasets
plotly
Pillow
numpy
openpyxl